chore: increase redis memory limit
As the current in-memory size is close to the limit, writing RDB
snapshots starts to fail with the following message:

  384:C 30 Sep 18:44:11.315 # Write error saving DB on disk: Out of memory

To accommodate for this, we raise the upper limit for now. We have a lot
of available memory in the cluster, so spiking a bit sometimes in order
to save the snapshots should not be problematic.

Not sure why redis is using this much memory; likely candidates are
large and/or many memote reports, or the fact that Celery requires that
much data. This should potentially be investigated in the future.